Move the deprecated code into its own file.
